move pickers to own folder in details, to make the component file structure a bit better
