+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/picker/datetime-picker.js
@@ -1,23 +1,23 @@
 import angular from "angular";
-import { attach, delay } from "../../utils.js";
-import { DomCache } from "../../cstm-ng-utils.js";
+import { attach, delay } from "../../../utils.js";
+import { DomCache } from "../../../cstm-ng-utils.js";
 
 const serviceDependencies = ["$scope", "$element"];
 function genComponentConf() {
   let template = `
-      <div class="datep__input">
-         <svg class="datep__input__icon clickable"
+      <div class="datetimep__input">
+         <svg class="datetimep__input__icon clickable"
             style="--local-primary:var(--won-primary-color);"
             ng-if="self.showResetButton"
-            ng-click="self.resetDate()">
+            ng-click="self.resetDatetime()">
             <use xlink:href="#ico36_close" href="#ico36_close"></use>
           </svg>
           <input
-              type="date"
-              class="datep__input__inner won-txt"
+              type="datetime-local"
+              class="datetimep__input__inner won-txt"
               placeholder="{{self.detail.placeholder}}"
               ng-model="self.value"
-              ng-change="::self.updateDate()"/>
+              ng-change="::self.updateDatetime()"/>
       </div>
     `;
 
@@ -26,27 +26,27 @@ function genComponentConf() {
       attach(this, serviceDependencies, arguments);
       this.domCache = new DomCache(this.$element);
 
-      window.datep4dbg = this;
+      window.datetimep4dbg = this;
 
-      this.addedDate = this.initialValue;
+      this.addedDatetime = this.initialValue;
       this.showResetButton = false;
 
-      delay(0).then(() => this.showInitialDate());
+      delay(0).then(() => this.showInitialDatetime());
     }
 
     /**
      * Checks validity and uses callback method
      */
-    update(date) {
-      if (date) {
-        this.onUpdate({ value: date });
+    update(datetime) {
+      if (datetime) {
+        this.onUpdate({ value: datetime });
       } else {
         this.onUpdate({ value: undefined });
       }
     }
 
-    showInitialDate() {
-      this.addedDate = this.initialValue;
+    showInitialDatetime() {
+      this.addedDatetime = this.initialValue;
 
       if (this.initialValue && this.initialValue.trim().length > 0) {
         this.textfield().value = this.initialValue.trim();
@@ -56,43 +56,43 @@ function genComponentConf() {
       this.$scope.$apply();
     }
 
-    updateDate() {
+    updateDatetime() {
       const text = this.textfield().value;
 
       if (text && text.trim().length > 0) {
-        this.addedDate = text.trim();
-        this.update(this.addedDate);
+        this.addedDatetime = text.trim();
+        this.update(this.addedDatetime);
         this.showResetButton = true;
       } else {
-        this.resetDate();
+        this.resetDatetime();
       }
     }
 
-    resetDate() {
-      this.addedDate = undefined;
+    resetDatetime() {
+      this.addedDatetime = undefined;
       this.textfield().value = "";
       this.update(undefined);
       this.showResetButton = false;
     }
 
     // textfieldNg() {
-    //   return this.domCache.ng(".datep__input__inner");
+    //   return this.domCache.ng(".datetimep__input__inner");
     // }
 
     // textfield() {
-    //   return this.domCache.dom(".datep__input__inner");
+    //   return this.domCache.dom(".datetimep__input__inner");
     // } // TODO: why is this done differently for number than for any other picker?
 
     textfieldNg() {
       return angular.element(this.textfield());
     }
     textfield() {
-      if (!this._dateInput) {
-        this._dateInput = this.$element[0].querySelector(
-          ".datep__input__inner"
+      if (!this._datetimeInput) {
+        this._datetimeInput = this.$element[0].querySelector(
+          ".datetimep__input__inner"
         );
       }
-      return this._dateInput;
+      return this._datetimeInput;
     }
   }
   Controller.$inject = serviceDependencies;
@@ -112,5 +112,5 @@ function genComponentConf() {
 }
 
 export default angular
-  .module("won.owner.components.datePicker", [])
-  .directive("wonDatePicker", genComponentConf).name;
+  .module("won.owner.components.datetimePicker", [])
+  .directive("wonDatetimePicker", genComponentConf).name;
