+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/picker/month-picker.js
@@ -1,23 +1,23 @@
 import angular from "angular";
-import { attach, delay } from "../../utils.js";
-import { DomCache } from "../../cstm-ng-utils.js";
+import { attach, delay } from "../../../utils.js";
+import { DomCache } from "../../../cstm-ng-utils.js";
 
 const serviceDependencies = ["$scope", "$element"];
 function genComponentConf() {
   let template = `
-      <div class="datetimep__input">
-         <svg class="datetimep__input__icon clickable"
+      <div class="monthp__input">
+         <svg class="monthp__input__icon clickable"
             style="--local-primary:var(--won-primary-color);"
             ng-if="self.showResetButton"
             ng-click="self.resetDatetime()">
             <use xlink:href="#ico36_close" href="#ico36_close"></use>
           </svg>
           <input
-              type="datetime-local"
-              class="datetimep__input__inner won-txt"
+              type="month"
+              class="monthp__input__inner won-txt"
               placeholder="{{self.detail.placeholder}}"
               ng-model="self.value"
-              ng-change="::self.updateDatetime()"/>
+              ng-change="::self.updateMonth()"/>
       </div>
     `;
 
@@ -26,27 +26,27 @@ function genComponentConf() {
       attach(this, serviceDependencies, arguments);
       this.domCache = new DomCache(this.$element);
 
-      window.datetimep4dbg = this;
+      window.monthp4dbg = this;
 
-      this.addedDatetime = this.initialValue;
+      this.addedMonth = this.initialValue;
       this.showResetButton = false;
 
-      delay(0).then(() => this.showInitialDatetime());
+      delay(0).then(() => this.showInitialMonth());
     }
 
     /**
      * Checks validity and uses callback method
      */
-    update(datetime) {
-      if (datetime) {
-        this.onUpdate({ value: datetime });
+    update(month) {
+      if (month) {
+        this.onUpdate({ value: month });
       } else {
         this.onUpdate({ value: undefined });
       }
     }
 
-    showInitialDatetime() {
-      this.addedDatetime = this.initialValue;
+    showInitialMonth() {
+      this.addedMonth = this.initialValue;
 
       if (this.initialValue && this.initialValue.trim().length > 0) {
         this.textfield().value = this.initialValue.trim();
@@ -56,43 +56,43 @@ function genComponentConf() {
       this.$scope.$apply();
     }
 
-    updateDatetime() {
+    updateMonth() {
       const text = this.textfield().value;
 
       if (text && text.trim().length > 0) {
-        this.addedDatetime = text.trim();
-        this.update(this.addedDatetime);
+        this.addedMonth = text.trim();
+        this.update(this.addedMonth);
         this.showResetButton = true;
       } else {
-        this.resetDatetime();
+        this.resetMonth();
       }
     }
 
-    resetDatetime() {
-      this.addedDatetime = undefined;
+    resetMonth() {
+      this.addedMonth = undefined;
       this.textfield().value = "";
       this.update(undefined);
       this.showResetButton = false;
     }
 
     // textfieldNg() {
-    //   return this.domCache.ng(".datetimep__input__inner");
+    //   return this.domCache.ng(".monthp__input__inner");
     // }
 
     // textfield() {
-    //   return this.domCache.dom(".datetimep__input__inner");
+    //   return this.domCache.dom(".monthp__input__inner");
     // } // TODO: why is this done differently for number than for any other picker?
 
     textfieldNg() {
       return angular.element(this.textfield());
     }
     textfield() {
-      if (!this._datetimeInput) {
-        this._datetimeInput = this.$element[0].querySelector(
-          ".datetimep__input__inner"
+      if (!this._monthInput) {
+        this._monthInput = this.$element[0].querySelector(
+          ".monthp__input__inner"
         );
       }
-      return this._datetimeInput;
+      return this._monthInput;
     }
   }
   Controller.$inject = serviceDependencies;
@@ -112,5 +112,5 @@ function genComponentConf() {
 }
 
 export default angular
-  .module("won.owner.components.datetimePicker", [])
-  .directive("wonDatetimePicker", genComponentConf).name;
+  .module("won.owner.components.monthPicker", [])
+  .directive("wonMonthPicker", genComponentConf).name;
